Move expression utils to util module
