+++ minicute-optimizer/lib/Minicute/Transpilers/Optimizers/ImmediateMatch.hs
@@ -5,13 +5,12 @@ module Minicute.Transpilers.Optimizers.ImmediateMatch
   ) where
 
 import Control.Lens.Each
-import Control.Lens.Getter ( to )
 import Control.Lens.Operators
 import Control.Lens.Plated ( transformOf )
 import Control.Lens.Wrapped ( _Wrapped )
 import Data.Data.Lens ( uniplate )
-import Data.List
 import Minicute.Data.Minicute.Program
+import Minicute.Utils.Minicute.Expression
 
 -- |
 -- An optimizer to remove immediate matches in a whole program.
@@ -25,7 +24,7 @@ immediateMatchMainEMC = transformOf uniplate go
   where
     go (ELet flag lDefs (EMatch (EVariable v) mCases))
       | Just vLDef <- lookupLDefsL v lDefs
-      , Just (tag, argExprs) <- destructDataExpression (vLDef ^. _letDefinitionBody)
+      , Just (tag, argExprs) <- destructStructureExpression (vLDef ^. _letDefinitionBody)
       , Just vMCase <- lookupMCasesL tag mCases
       = let
           argBinders = vMCase ^. _matchCaseArguments
@@ -39,23 +38,3 @@ immediateMatchMainEMC = transformOf uniplate go
         in
           ELet flag lDefs expr
     go e = e
-
--- |
--- __TODO: move this into a Util or Data module__
-destructDataExpression :: ExpressionMC a -> Maybe (Integer, [ExpressionMC a])
-destructDataExpression e = go e []
-  where
-    go (EConstructor tag arity) args
-      | arity == genericLength args = Just (tag, args)
-    go (EApplication e1 e2) args = go e1 (e2 : args)
-    go _ _ = Nothing
-
--- |
--- __TODO: move this into a Util or Data module__
-lookupMCasesL :: Integer -> [MatchCase ExpressionMC a] -> Maybe (MatchCase ExpressionMC a)
-lookupMCasesL tag = find (^. _matchCaseTag . to (== tag))
-
--- |
--- __TODO: move this into a Util or Data module__
-lookupLDefsL :: (Eq a) => a -> [LetDefinition ExpressionMC a] -> Maybe (LetDefinition ExpressionMC a)
-lookupLDefsL binder = find (^. _letDefinitionBinder . to (== binder))
